Implemented MaxSpeed for all GameObjects, PlayerShip stops at viewport boundaries, PlayerShip acceleration reduced and MaxSpeed set.
For all GameObjects, whenever Speed is set it is limited to MaxSpeed, which is initialized to an arbitrariy high value by default, MaxSpeed needs to be specifically set to have any noticeable effect. PlayerShip now has a MaxSpeed.
PlayerShip checks against viewport height and width whenever it is updated; if it hits a viewport boundary, position is limited to that boundary (minus a small constant to keep the entire ship within view) and velocity in that direction is set to zero.
PlayerShip acceleration coefficient reduced from 0.1f to 0.03f, MaxSpeed set to 0.8f.